pagination skeleton with a lot of broken code
